requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
tensorflow==2.13.0
shap==0.42.1
lime==0.2.0.1
pillow==10.0.0
opencv-python==4.8.0.76
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1